Disable Style/CaseEquality
In general, I agree that `===` should be discouraged (on `Enumerable`,
`Regexp`, etc), but I don't think we can disallow it using Rubocop, and
in at least one scenario it should actually be the encouraged way of
doing things: hierarchy membership checks (aka: typechecking, but not
quite).

The idiomatic way of doing hierarchy membership checking is typically
using the `is_a?` or `kind_of?`, but this should actually be
discouraged, for two reasons.

1. `BasicObject` doesn't respond to `is_a?` and so whenever we use `is_a?`.
  When we use `is_a?` we introduce coupling (which is generally
  unnecessary), between the code that checks and the code that produces
  the objects. This practice is also prevented by Sorbet when using
  interfaces, because the implementation could very well be a
  `BasicObject`. In the Checkouts component, we've had to disable the
  check entirely since we use Sorbet with strict typing enabled.

2. When checking membership, we should always ask the collection
  rather than the item. E.g., we do `array.include?(item)` and never
  `item.contained_in?(array)`.

Because Rubocop doesn't have the required knowledge to know the type of
the receiver, the only option we have is to disable this cop.